let notification settings to control all distribution channels, not only push but also in-site notifications
